fix(web): cancel inactive message backfills (#2287)
